Uploaded filess for GameLogic and Parser
GameLogic file for creating the game, Parser to handle text responses, Command for handling actions based on Parser input, and Room to allow for GameLogic to proper create the game based on scenario/room files. Still needs a little work to properly mesh with rest of code, but something for everyone to refer to.